Add copyright to Charlie Waters from item prices
I'm not honestly sure if this is the right thing to do. I think it is
for Overlay because it's largely their code, trimmed down, with a tiny bit
of new I wrote. But for config and plugin, it's all new I wrote or
edited from ExamplePlugin, but it's so similar (since I used his code
as a reference to see how to write a RuneLite plugin in general), that
I think it might be right to credit them there too.